Split unit tests for create command.
